map coordinates fix, current medication price fix

+++ backend/src/main/java/app/model/user/Address.java
@@ -8,8 +8,8 @@ public class Address {
     private String street;
     private String town;
     private String country;
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
 
     public Address() {
     }
@@ -38,19 +38,19 @@ public class Address {
         this.country = country;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
